refactor: prepare for more dab functions apart of NFTs
