+++ src/dab_canisters/index.ts
@@ -1,0 +1,1 @@
+export * from './nfts_registry';
